# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SAHU CSC FV1 - CSC Business Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    patch:
      operationId: updateProfile
      tags: [profile]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileInput"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /profile/avatar:
    post:
      operationId: uploadAvatar
      tags: [profile]
      summary: Upload or replace profile picture (base64 data URL)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AvatarInput"
      responses:
        "200":
          description: Avatar uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvatarResponse"
    delete:
      operationId: deleteAvatar
      tags: [profile]
      summary: Remove profile picture
      responses:
        "200":
          description: Avatar removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # Preferences
...
  /reports/daily:
    get:
      operationId: getDailyReport
      tags: [reports]
      summary: Get daily report
      parameters:
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Daily report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyReport"
  /reports/monthly:
    get:
      operationId: getMonthlyReport
      tags: [reports]
      summary: Get monthly report
      parameters:
        - name: year
          in: query
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Monthly report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlyReport"
  /reports/service-breakdown:
    get:
      operationId: getServiceBreakdown
      tags: [reports]
      summary: Get revenue breakdown by service type
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Service breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceBreakdown"
  /reports/export:
    get:
      operationId: exportReport
      tags: [reports]
      summary: Export report to Excel
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [daily, monthly, custom]
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Excel file
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary

  # Notifications
...
components:
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # Auth
...
    UserProfile:
      type: object
      required: [id, username, email, role]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        mobile:
          type: ["string", "null"]
        fullName:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, operator, user]
        profilePicture:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        createdAt:
          type: string
    UpdateProfileInput:
      type: object
      properties:
        fullName:
          type: string
        email:
          type: string
        mobile:
          type: string
        bio:
          type: string
        address:
          type: string
        password:
          type: string
        currentPassword:
          type: string
    AvatarInput:
      type: object
      required: [profilePicture]
      properties:
        profilePicture:
          type: string
          description: Base64 data URL (data:image/...;base64,...)
    AvatarResponse:
      type: object
      required: [profilePicture]
      properties:
        profilePicture:
          type: ["string", "null"]
...
    DailyReport:
      type: object
      required: [date, transactionCount, totalCredits, totalDebits, netRevenue, topServices]
      properties:
        date:
          type: string
        transactionCount:
          type: integer
        totalCredits:
          type: number
        totalDebits:
          type: number
        netRevenue:
          type: number
        topServices:
          type: array
          items:
            $ref: "#/components/schemas/ServiceBreakdown"
    MonthlyReport:
      type: object
      required: [year, month, totalTransactions, totalCredits, totalDebits, netProfit, dailyBreakdown, topServices]
      properties:
        year:
          type: integer
        month:
          type: integer
        totalTransactions:
          type: integer
        totalCredits:
          type: number
        totalDebits:
          type: number
        netProfit:
          type: number
        dailyBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/DailyBreakdown"
        topServices:
          type: array
          items:
            $ref: "#/components/schemas/ServiceBreakdown"
    DailyBreakdown:
      type: object
      required: [date, credits, debits, transactions]
      properties:
        date:
          type: string
        credits:
          type: number
        debits:
          type: number
        transactions:
          type: integer
    ServiceBreakdown:
      type: object
      required: [serviceType, count, revenue]
      properties:
        serviceType:
          type: string
        count:
          type: integer
        revenue:
          type: number

    # Notifications
...