# lib/api-zod/src/generated/types/avatarInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SAHU CSC FV1 - CSC Business Management Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AvatarInput {
  /** Base64 data URL (data:image/...;base64,...) */
  profilePicture: string;
}