# lib/api-zod/src/generated/types/avatarResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SAHU CSC FV1 - CSC Business Management Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AvatarResponse {
  /** @nullable */
  profilePicture: string | null;
}