# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SAHU CSC FV1 - CSC Business Management Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserProfileRole } from './userProfileRole';

export interface UserProfile {
  id: number;
  username: string;
  email: string;
  /** @nullable */
  mobile?: string | null;
  /** @nullable */
  fullName?: string | null;
  role: UserProfileRole;
  /** @nullable */
  profilePicture?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  address?: string | null;
  createdAt?: string;
}